requirements.txt
uvicorn~=0.34.3
fastapi~=0.115.12
pytest~=8.4.0
requests~=2.32.3
python-dotenv~=1.1.0